pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crawlstudio"
version = "0.1.0"
authors = [
  { name = "Saeed Ashraf", email = "saeedashraf@example.com" },
]
description = "Unified Python library for web crawling tools - supports Firecrawl, Crawl4AI, Scrapy, and Browser-Use backends"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
keywords = ["web scraping", "crawling", "firecrawl", "scrapy", "crawl4ai", "browser automation"]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0", 
    "python-dotenv>=1.0",
    "crawl4ai>=0.7.4",
    "firecrawl-py>=0.0.20",
    "scrapy>=2.11",
    "aiohttp>=3.9",
    "diskcache>=5.6",
]

[project.optional-dependencies]
browser-use = [
    "browser-use",
    "langchain-openai",
    "langchain-anthropic",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/saeedashraf/CrawlStudio"
Repository = "https://github.com/saeedashraf/CrawlStudio"
Issues = "https://github.com/saeedashraf/CrawlStudio/issues"
Documentation = "https://github.com/saeedashraf/CrawlStudio#readme"

[project.scripts]
crawlstudio = "crawlstudio.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["crawlstudio"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true